daily_stoic_quote/requirements.txt
fastapi
uvicorn
requests
python-dotenv
Jinja2